Test and QA targets for the Makefile.
This also defines part of the style for the project. Formatting
is per rustfmt except that lines are limited to 79 characters
instead of 100. Linting is done with clippy in pedantic mode.
Testing... runs the zero tests that currently exist.

The idea is: Travis CI is the test runner of record. Pull
requests will be required to pass all checks in Travis before
being merged, including lint and formatting checks. `make check`
runs the Travis checks locally (as well as it can).

Note that test coverage measurement is missing. Apparently this
is a little tricky (particularly on macOS) so I'm not going to
worry about it for now. But it's something that should be
circled back to at some point.