Major invocation infrastructure rewrite to support signals.
The main thread is now used exclusively for listening for
control events: signals from the OS or messages from work
threads indicating their completion.

The first INT or TERM signal sets a flag on the `Invocation`
instance that commands can check with the `should_terminate()`
method. When that flag is `true` the command should clean up
and return as soon as it can.

If the main thread gets another INT or TERM signal before the
command has sent its "complete" message, the program simply
terminates then and there.

(Termination signals are also handled during `app::init()`. In
that case, there's nothing to clean up so the program simply
exits with an exit code of 1.)