Implements graceful termination of the pull command.
Since it's unclear to me whether `git fetch` is safe to kill
(and there are enough StackOverflow questions about corrupted
repos after borked fetches to lead me to believe it's not) the
model is to let the running fetches complete and "cancel" the
rest of the pending fetches.

The rough infrastructure is there to implement a hard kill on
a second sigterm/ctrl-c, but this is not yet implemented.