Removes paging for mgit output.
The basic problem is signal handling. We're interested in SIGTERM and SIGINT,
which mgit takes as a signal from the user to stop the program. Specifically,
we're concerned during `mgit pull` and `mgit status`. (`mgit status` is
normally a very fast command, but can get hung up for ~seconds if it has to
report on the 20k untracked files in the `node_modules/` directory you forgot
to add to `.gitignore`. ~Seconds is plenty of time for the user to hit ctrl-c,
and mgit needs to do the right thing when that happens.)

Currently, paging is implemented like this:

- fork the process
- hook up child stdout to parent stdin
- execvp `less -options` in the parent
- let the child return to the calling function to continue execution

When the user hits ctrl-c in the terminal, the signal gets sent to the parent
process (the pager program). The child process never sees the signal so,
depending on the exact sequence of events, the user experience ranges from
unexpected (partial output) to downright broken (the child panics when it can't
write to stdout because the pager program has exited and closed the pipe).

(Note that using the `kill` command from another shell on the pid of the child
process would work as expected, but who cares? Ctrl-c is what we're really
interested in.)

I pulled the code from the pager crate into the tree and tried a number of
things that didn't work:

1. Masking signals in the parent. This just had the effect of making the
program completely unkillable. The child of the fork (where we can sanely
handle signals) still wasn't getting signals.

2. Registering a signal handler before execvp. I didn't think this would
actually work, and it didn't. I assume it's because any signal handlers
registered by the execvp-ed program get to do their thing before (or instead
of) ours. (Or it could be completely invalid -- the time between `fork` and
`exec` is weird and signals are weird, and it wouldn't surprise me if those
things just don't play nicely in the way I was putting them together.)

3. Swapping the parent and child. Again, intuitively this shouldn't work and it
didn't, though it was the closest. This allows for masking of signals in the
child (pager) while handling them in the parent (mgit) -- great! The problem
is, even after closing the write side of the pipe in the parent, the child
process never seems to see EOF on the read side so it never exits from the
pager, meaning the parent hangs forever on `wait()`.

Honestly, I'm not even sure that last bit is an accurate assessment. I don't
understand exactly how the pager, as the child, is getting/processing
keystrokes (you can page through to the end of the output using space). And I
don't understand why closing three of the four pipe references (write side in
the child, read side then write side in the parent) doesn't cause the read side
in the child to get an EOF which, in conjunction with the `-E` argument passed
to `less`, should make the child process exit.

Taking a step back, it's not obvious to me that -- even if it worked -- the
complexity is worth the feature. The code for implementing (3) was write-once,
never-touch-again. (And definitely not cross-platform.) Not the kind of
liability I'm looking for with mgit.

So the choice is obvious. mgit can live without a pager, mgit cannot live
without properly handling signals. I'm not smart enough to have both, so the
pager has to go.