Formats comments consistently.
The tool used for formatting is in tool/comments. Run with `-c`
to output a diff instead of rewriting files. Note that it's a
really crappy tool. (Automatically formatting comments is a
tough job due to things like bulleted lists and code samples.
To do it "right" you'd need to have the parsed markup tree.)
The only reason it's included is because it might be just useful
enough until comment formatting has landed in rustfmt.

That said, the idea is to wrap prose-like comments to a maximum
width of 79 characters per line of text, not to exceeed 99
characters when considering indentation and comment prefix.
That is, a comment with little or no indentation (e.g. a header
comment) will wrap at, say, column 83 while a comment in a
deeply-nested loop will wrap at column 99. This (hopefully)
avoids lines that are uncomfortably long to read and makes
rustfmt happy while making the best use of the space available
in the source files.